+++ jest.config.js
@@ -160,7 +160,7 @@ module.exports = {
   testRegex: [
     // "base\.test\.js",
     // "other\.test\.js",
-    "0.5",
+    // "beforeV0.5",
   ],
 
   // This option allows the use of a custom results processor