+++ rollup.prod.ts
@@ -1,38 +1,59 @@
 import pkg from "./package.json"
 import typescript from "@rollup/plugin-typescript"
 import del from "rollup-plugin-delete"
-import { uglify } from "rollup-plugin-uglify"
+import { terser } from "rollup-plugin-terser"
+import { getBabelOutputPlugin } from "@rollup/plugin-babel"
 
 export default [{
-    // treeshake: false,
     input: './src/index.ts',
-    output: [{//for umd publish
-        file: pkg.main,
-        format: 'umd',
-        name: pkg.name,
-    }, {//for esm publish
+    output: {// for esm publish
         file: pkg.module,
         format: 'esm',
         name: pkg.name,
-    }, {//for test
-        file: pkg.main.replace('bin', 'test/src').replace('umd', 'iife'),
-        format: 'iife',
-        // name for testEnv puppeteer
-        name: 'test',
-        plugins: [
-            del({ targets: ['./test/src/types'], hook: 'writeBundle' })
-        ]
-    }],
+    },
     plugins: [
-        del({
-            targets: ['./bin', './test/src'],
+        del({ targets: ['./bin'], }),
+        del({ targets: ['./bin/types/pipes.d.ts'], hook: 'closeBundle' }),
+        del({ targets: ['./bin/types/polyfill'], hook: 'closeBundle' }),
+        typescript({
+            tsconfig: './tsconfig.json',
+            declaration: true,
+            declarationDir: "./types",
         }),
+        getBabelOutputPlugin({
+            presets: [['@babel/preset-env', {
+                useBuiltIns: 'usage', corejs: { version: '3.15' },
+                targets: pkg.browserslist,
+            },]],
+            plugins: [['@babel/plugin-transform-runtime', { useESModules: true }]]
+        }),
+        terser(),
+    ],
+    external: [/@babel\/runtime/],
+}, {
+    input: './src/index.ts',
+    output: {// for umd publish
+        file: pkg.main,
+        // format: 'umd',
+        name: pkg.name,
+    },
+    plugins: [
         typescript({
             tsconfig: './tsconfig.json',
         }),
-        uglify(),
+        getBabelOutputPlugin({
+            presets: [['@babel/preset-env', {
+                modules: 'umd',
+                useBuiltIns: 'usage', corejs: { version: '3.15' },
+                targets: pkg.browserslist,
+            }]],
+            plugins: [['@babel/plugin-transform-runtime', { useESModules: false }]]
+        }),
+        terser(),
     ],
-}, {// for polyfill
+    external: [/@babel\/runtime/],
+},
+{// for polyfill
     input: {
         globalThis: './src/polyfill/globalThis.ts',
     },
@@ -47,7 +68,22 @@ export default [{
             tsconfig: false,
             include: './src/polyfill/*',
         }),
-        uglify(),
+        terser(),
+    ]
+}, {// for browser test
+    input: './src/index.ts',
+    output: {
+        file: pkg.main.replace('bin', 'test/src').replace('umd', 'iife'),
+        format: 'iife',
+        // name for testEnv puppeteer
+        name: 'test',
+    },
+    plugins: [
+        del({ targets: ['./test/src'], }),
+        typescript({
+            tsconfig: './tsconfig.json',
+        }),
+        terser(),
     ]
 }, {// common pipes for test
     input: './src/pipes.ts',
@@ -58,8 +94,6 @@ export default [{
     },
     plugins: [
         del({ targets: ['./test/pipes.js'] }),
-        typescript({
-            tsconfig: false,
-        })
+        typescript({ tsconfig: false, target: "ESNext" }),
     ]
 }]