+++ test/beforeV0.5.test.js
@@ -9,15 +9,15 @@ describe('Before version 0.5', () => {
             }, {
                 urnStringTest: 'test',
                 customConfigTest: {
-                    urn: 'test', config: { custom: { customTag: true } }
+                    urn: 'customConfigTest', config: { custom: { customTag: true } }
                 },
             })
         })
         await page.addScriptTag({ content: script })
     })
 
-    describe('URN type string test', () => {
-        test('URN type string', async () => {
+    describe('APIConfig type URN test', () => {
+        test('APIConfig type string', async () => {
             const _ = await page.evaluate(async () => {
                 const final = {}
                 const _ = Api.request.use((url, config) => {
@@ -38,6 +38,7 @@ describe('Before version 0.5', () => {
                 url: 'https://www.example.com/test'
             })
         })
+        test.todo('APIConfig type function')
     })
 
     describe('Custom config test', () => {
@@ -49,12 +50,73 @@ describe('Before version 0.5', () => {
                 }, pipes.NoRequest)
                 try {
                     await Api.customConfigTest()
-                } catch (e) {
-                }
+                } catch (e) { }
                 Api.request.eject(_)
                 return final
             })
             expect(_).toBeTruthy()
         })
     })
+
+    describe('Async PipeRequest', () => {
+        test('Async PipeRequest Delay', async () => {
+            const _ = await page.evaluate(async () => {
+                let t1, t2
+                const _ = Api.request.use(() => { t1 = Date.now() },
+                    pipes.RequestDelay(1000),
+                    () => { t2 = Date.now() },
+                    pipes.NoRequest)
+                try {
+                    await Api.customConfigTest()
+                } catch (e) { }
+                Api.request.eject(_)
+                return t2 - t1 > 1000
+            })
+            expect(_).toBeTruthy()
+        })
+    })
+
+    describe('Dynamic config', () => {
+        test('Dynamic config', async () => {
+            const _ = await page.evaluate(async () => {
+                let final = {}
+                const _ = Api.request.use((url, config) => {
+                    final.dynamic = config?.custom?.customTag
+                    final.method = config?.method
+                }, pipes.NoRequest)
+                try {
+                    await Api.urnStringTest(null, null, { method: 'POST', custom: { customTag: true } })
+                } catch (e) { }
+                Api.request.eject(_)
+                return final
+            })
+            expect(_).toEqual({
+                dynamic: true,
+                method: 'POST'
+            })
+        })
+        test('Config custom with object.assign', async () => {
+            const _ = await page.evaluate(async () => {
+                let final = {}
+                const _ = Api.request.use((url, config) => {
+                    final.customTag = config?.custom?.customTag
+                    final.tag = config?.custom?.tag
+                }, pipes.NoRequest)
+                try {
+                    await Api.customConfigTest(null, null, { custom: { customTag: false, tag: true } })
+                } catch (e) { }
+                Api.request.eject(_)
+                return final
+            })
+            expect(_).toEqual({
+                customTag: false,
+                tag: true
+            })
+        })
+    })
+
+    describe('Params parse to url.search', () => {
+        test.todo('next version')
+    })
+
 })