+++ test/pipes.js
@@ -1,19 +1,21 @@
 var pipes = (function (exports) {
     'use strict';
 
-    var NoRequest = function (url, config, _a, _b) {
-        _a[0]; _a[1];
-        _b[0]; _b[1]; _b[2]; _b[3];
-        console.log("try fetch " + url + ", blocked by NoRepuest Pipe");
+    const NoRequest = (url, config, [body, params], [api, urn, config_, baseConfig]) => {
+        console.log(`try fetch ${url}, blocked by NoRepuest Pipe`);
         return 'No Request';
     };
-    var RequestLogger = function (url, config, _a, _b) {
-        var body = _a[0], params = _a[1];
-        var api = _b[0], urn = _b[1], config_ = _b[2], baseConfig = _b[3];
+    const RequestLogger = (url, config, [body, params], [api, urn, config_, baseConfig]) => {
+        console.log('RequestLogger begin');
         console.log(url.toString(), JSON.stringify(config), JSON.stringify([body, params]), JSON.stringify([api, urn, config_, baseConfig]));
+        console.log('RequestLogger end');
+    };
+    const RequestDelay = (delay) => async () => {
+        await new Promise(resolve => setTimeout(resolve, delay));
     };
 
     exports.NoRequest = NoRequest;
+    exports.RequestDelay = RequestDelay;
     exports.RequestLogger = RequestLogger;
 
     Object.defineProperty(exports, '__esModule', { value: true });